added functionality and new module to scripts folder
Added new utils.py inside scripts folder, has a converSize() function
for converting between kB, Mb and so forth.

Used the new function to convert the output of getfreeMem() and
gettotalMem() to show a more human readable format.

made sure that getFacts() displays STR and not in unicode format, while
keeping the data in tact.