Allowed input of username and password
Added functionality to input username and password at the command-line.
Only works with eapi, similar method needs to be added to individual
platforms setLogin(self, username, password).

Also allowed for -f or -c to be used if either is entered, fancy
redundancy that was already in argparse, but not being taken advantage
of.